# gen/http/openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fire Tracker Service",
    "description": "Service for fire tracking",
    "termsOfService": "terms",
    "contact": {
      "name": "support",
      "url": "firetracker.com",
      "email": "support@firetracker.com"
    },
    "license": {
      "name": "GPT-3",
      "url": "licence url"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://test.tmf.com"
    },
    {
      "url": "https://trackmyfire.com"
    }
  ],
  "paths": {
...
    "/log": {
      "post": {
        "tags": [
          "log"
        ],
        "summary": "create log",
        "description": "Add a log to existing fire",
        "operationId": "log#create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestBody"
              },
              "example": {
                "addedAt": "2020-01-02T00:00:00Z",
                "createdAt": "2020-01-02T00:00:00Z",
                "fireID": 1,
                "id": 3719654112391162300,
                "name": "Big Log",
                "size": "M",
                "updatedAt": "2020-01-02T00:00:00Z",
                "weather": {
                  "createdAt": "2020-01-02T00:00:00Z",
                  "dewPoint": {
                    "unit": "K",
                    "value": 20
                  },
                  "fireID": 1,
                  "humidity": 1500,
                  "id": 1,
                  "logID": 1,
                  "temperature": {
                    "unit": "K",
                    "value": 20
                  },
                  "weatherTime": "2020-01-02T00:00:00Z",
                  "weatherType": "Sunny",
                  "wind": {
                    "direction": "SE",
                    "speed": 15,
                    "unit": "KPH"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRequestBody"
                },
                "example": {
                  "addedAt": "2020-01-02T00:00:00Z",
                  "createdAt": "2020-01-02T00:00:00Z",
                  "fireID": 1,
                  "id": 6083859619097906619,
                  "name": "Big Log",
                  "size": "M",
                  "updatedAt": "2020-01-02T00:00:00Z",
                  "weather": {
                    "createdAt": "2020-01-02T00:00:00Z",
                    "dewPoint": {
                      "unit": "C",
                      "value": 20
                    },
                    "fireID": 1,
                    "humidity": 1500,
                    "id": 1,
                    "logID": 1,
                    "temperature": {
                      "unit": "C",
                      "value": 20
                    },
                    "weatherTime": "2020-01-02T00:00:00Z",
                    "weatherType": "Sunny",
                    "wind": {
                      "direction": "E",
                      "speed": 15,
                      "unit": "MPH"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "id": "3F1FKVRR",
                  "message": "Value of ID must be an integer",
                  "name": "bad_request"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRequestBody"
                },
                "example": {
                  "addedAt": "2020-01-02T00:00:00Z",
                  "createdAt": "2020-01-02T00:00:00Z",
                  "fireID": 1,
                  "id": 84502070544293262,
                  "name": "Big Log",
                  "size": "XL",
                  "updatedAt": "2020-01-02T00:00:00Z",
                  "weather": {
                    "createdAt": "2020-01-02T00:00:00Z",
                    "dewPoint": {
                      "unit": "C",
                      "value": 20
                    },
                    "fireID": 1,
                    "humidity": 1500,
                    "id": 1,
                    "logID": 1,
                    "temperature": {
                      "unit": "C",
                      "value": 20
                    },
                    "weatherTime": "2020-01-02T00:00:00Z",
                    "weatherType": "Sunny",
                    "wind": {
                      "direction": "E",
                      "speed": 15,
                      "unit": "MPH"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateRequestBody": {
        "type": "object",
        "properties": {
          "addedAt": {
            "type": "string",
            "description": "time log was added to fire",
            "example": "2020-01-02T00:00:00Z",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "name",
            "example": "2020-01-02T00:00:00Z",
            "format": "date-time"
          },
          "fireID": {
            "type": "integer",
            "description": "Fire identifier log belongs to",
            "example": 1,
            "format": "int64"
          },
          "id": {
            "type": "integer",
            "description": "id",
            "example": 6647109503785378516,
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "name",
            "example": "Big Log"
          },
          "size": {
            "type": "string",
            "description": "size of log",
            "example": "L",
            "enum": [
              "S",
              "M",
              "L",
              "XL"
            ]
          },
          "updatedAt": {
            "type": "string",
            "description": "name",
            "example": "2020-01-02T00:00:00Z",
            "format": "date-time"
          },
          "weather": {
            "$ref": "#/components/schemas/Weather"
          }
        },
        "example": {
          "addedAt": "2020-01-02T00:00:00Z",
          "createdAt": "2020-01-02T00:00:00Z",
          "fireID": 1,
          "id": 5114972834516992081,
          "name": "Big Log",
          "size": "L",
          "updatedAt": "2020-01-02T00:00:00Z",
          "weather": {
            "createdAt": "2020-01-02T00:00:00Z",
            "dewPoint": {
              "unit": "K",
              "value": 20
            },
            "fireID": 1,
            "humidity": 1500,
            "id": 1,
            "logID": 1,
            "temperature": {
              "unit": "K",
              "value": 20
            },
            "weatherTime": "2020-01-02T00:00:00Z",
            "weatherType": "Sunny",
            "wind": {
              "direction": "SE",
              "speed": 15,
              "unit": "KPH"
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "fault": {
            "type": "boolean",
            "description": "Is the error a server-side fault?",
            "example": false
          },
          "id": {
            "type": "string",
            "description": "ID is a unique identifier for this particular occurrence of the problem.",
            "example": "123abc"
          },
          "message": {
            "type": "string",
            "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
            "example": "parameter 'p' must be an integer"
          },
          "name": {
            "type": "string",
            "description": "Name is the name of this class of errors.",
            "example": "bad_request"
          },
          "temporary": {
            "type": "boolean",
            "description": "Is the error temporary?",
            "example": true
          },
          "timeout": {
            "type": "boolean",
            "description": "Is the error a timeout?",
            "example": true
          }
        },
        "description": "Please ensure you are sending valid data.",
        "example": {
          "id": "3F1FKVRR",
          "message": "Value of ID must be an integer",
          "name": "bad_request"
        },
        "required": [
          "name",
          "id",
          "message",
          "temporary",
          "timeout",
          "fault"
        ]
      },
...
      "Temperature": {
        "type": "object",
        "properties": {
          "unit": {
            "type": "string",
            "description": "measurement unit",
            "example": "F",
            "enum": [
              "K",
              "C",
              "F"
            ]
          },
          "value": {
            "type": "integer",
            "description": "temperature value",
            "example": 20,
            "format": "int32"
          }
        },
        "example": {
          "unit": "F",
          "value": 20
        }
      },
...
      "Weather": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "description": "name",
            "example": "2020-01-02T00:00:00Z",
            "format": "date-time"
          },
          "dewPoint": {
            "$ref": "#/components/schemas/Temperature"
          },
          "fireID": {
            "type": "integer",
            "description": "ID of the fire",
            "example": 1,
            "format": "int64"
          },
          "humidity": {
            "type": "integer",
            "description": "humidity level",
            "example": 1500,
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "description": "id",
            "example": 1,
            "format": "int64"
          },
          "logID": {
            "type": "integer",
            "description": "ID of the log",
            "example": 1,
            "format": "int64"
          },
          "temperature": {
            "$ref": "#/components/schemas/Temperature"
          },
          "weatherTime": {
            "type": "string",
            "example": "2020-01-02T00:00:00Z",
            "format": "date-time"
          },
          "weatherType": {
            "type": "string",
            "description": "type of weather",
            "example": "Sunny",
            "enum": [
              "Sunny",
              "Cloudy",
              "Raining",
              "Windy"
            ]
          },
          "wind": {
            "$ref": "#/components/schemas/Wind"
          }
        },
        "example": {
          "createdAt": "2020-01-02T00:00:00Z",
          "dewPoint": {
            "unit": "K",
            "value": 20
          },
          "fireID": 1,
          "humidity": 1500,
          "id": 1,
          "logID": 1,
          "temperature": {
            "unit": "K",
            "value": 20
          },
          "weatherTime": "2020-01-02T00:00:00Z",
          "weatherType": "Windy",
          "wind": {
            "direction": "SE",
            "speed": 15,
            "unit": "KPH"
          }
        }
      },
...
      "Wind": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string",
            "description": "wind direction",
            "example": "S",
            "enum": [
              "S",
              "SE",
              "E",
              "NE",
              "N",
              "NW",
              "W",
              "SW"
            ]
          },
          "speed": {
            "type": "integer",
            "description": "wind speed",
            "example": 15,
            "format": "int32"
          },
          "unit": {
            "type": "string",
            "description": "measurement unit",
            "example": "KPH",
            "enum": [
              "KPH",
              "MPH"
            ]
          }
        },
        "example": {
          "direction": "NE",
          "speed": 15,
          "unit": "MPH"
        }
      }
    }
  }
}